Having proplems with then Airplane db

+++ AirportSystem/Models/Airplane.cs
@@ -9,20 +9,20 @@
 
 namespace AirportSystem.Models
 {
-    using System;
-    using System.Collections.Generic;
-    
-    public partial class Airplane
+	using System;
+	using System.Collections.Generic;
+	using System.ComponentModel.DataAnnotations;
+
+	public partial class Airplane
     {
         [System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Usage", "CA2214:DoNotCallOverridableMethodsInConstructors")]
         public Airplane()
         {
             this.Schedule = new HashSet<Schedule>();
         }
-    
         public string RegNumber { get; set; }
-        public short Capacity { get; set; }
-        public double Length { get; set; }
+        public int Capacity { get; set; }
+        public int Length { get; set; }
         public int AirplaneTypesId { get; set; }
     
         public virtual AirplaneType AirplaneTypes { get; set; }
